# perf: optimize CBLSSignature::VerifySecureAggregated in 2 times