Current implementation of `CoreMPL::VerifySecure` does serialization and deserialization of G1 element.
    
Deserialization is very heavy because FromBytes() also validates that G1 is valid.
This validation is very expensive for computation power but useless because we already know that object is valid.
It is computation expensive. Much better to keep pointer to original object.
    
It speeds up validation of quorum commitment in Dash Core in 2 times which make indexing and validating new blocks 11% faster.
The performance improvement is most noticeable on blocks that have rotation quorum commitment, because one block has 32 commitments. 


Benchmarks has been done on 15000 last blocks.

Dash Core with this PR:
<img width="695" alt="image" src="https://github.com/user-attachments/assets/ca6439cd-aef3-4c22-be1a-423b33376544" />

```
2025-05-26T07:35:21Z [bench]         - m_qblockman: 0.05ms [55.43s]
2025-05-26T07:35:21Z [bench]   - Connect total: 19.96ms [223.70s (14.40ms/blk)]
2025-05-26T07:35:21Z [bench] - Connect block: 20.46ms [230.66s (14.85ms/blk)]
```


Dash Core with current master:
<img width="695" alt="image" src="https://github.com/user-attachments/assets/4708ebd4-e17b-4e8d-bf67-ebc1d4861fca" />

```
2025-05-22T18:39:44Z [bench]         - m_qblockman: 0.03ms [96.90s]
2025-05-22T18:39:44Z [bench]   - Connect total: 9.31ms [252.80s (16.28ms/blk)]
2025-05-22T18:39:44Z [bench] - Connect block: 9.50ms [258.90s (16.67ms/blk)]
```